+++ libs/newsletter-workflow/src/lib/getInitialStateForLaunch.ts
